+++ WaveletStatistics/WaveletStatistics.py
@@ -137,7 +137,7 @@ numpyNipyDataSz = numpy.asarray(numpyNipyData.shape)
 
 
 #set up bool for keeping middle slices...
-nElimSlices = round(numpyNipyDataSz[2] * fracElimSlices)
+nElimSlices = int(round(numpyNipyDataSz[2] * fracElimSlices))
 boolElimSlices = numpy.zeros(numpyNipyDataSz[2], dtype=numpy.int)
 if (fracElimSlicesUse == 'y'):
     boolElimSlices[nElimSlices:(numpyNipyDataSz[2] - nElimSlices)] = 1
